Module google_runtimeconfig1_beta1::api

Structs

Binding
Associates members with a role.

Cardinality
A Cardinality condition for the Waiter resource. A cardinality condition is met when the number of variables under a specified path prefix reaches a predefined number. For example, if you set a Cardinality condition where the path is set to /foo and the number of paths is set to 2, the following variables would meet the condition in a RuntimeConfig resource: + /foo/variable1 = "value1" + /foo/variable2 = "value2" + /bar/variable3 = "value3" It would not satisfy the same condition with the number set to 3, however, because there is only 2 paths that start with /foo. Cardinality conditions are recursive; all subtrees under the specific path prefix are counted.

CloudRuntimeConfig
Central instance to access all CloudRuntimeConfig related resource activities

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

EndCondition
The condition that a Waiter resource is waiting for.

Expr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec. Example (Comparison): title: "Summary size limit" description: "Determines if a summary is less than 100 chars" expression: "document.summary.size() < 100" Example (Equality): title: "Requestor is owner" description: "Determines if requestor is the document owner" expression: "document.owner == request.auth.claims.email" Example (Logic): title: "Public documents" description: "Determine whether the document should be publicly visible" expression: "document.type != 'private' && document.type != 'internal'" Example (Data Manipulation): title: "Notification string" description: "Create a notification string with a timestamp." expression: "'New message received at ' + string(document.create_time)" The exact variables and functions that may be referenced within an expression are determined by the service that evaluates it. See the service documentation for additional information.

ListConfigsResponse
ListConfigs() returns the following response. The order of returned objects is arbitrary; that is, it is not ordered in any particular way.

ListVariablesResponse
Response for the ListVariables() method.

ListWaitersResponse
Response for the ListWaiters() method. Order of returned waiter objects is arbitrary.

Operation
This resource represents a long-running operation that is the result of a network API call.

Policy
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources. A Policy is a collection of bindings. A binding binds one or more members to a single role. Members can be user accounts, service accounts, Google groups, and domains (such as G Suite). A role is a named list of permissions; each role can be an IAM predefined role or a user-created custom role. For some types of Google Cloud resources, a binding can also specify a condition, which is a logical expression that allows access to a resource only if the expression evaluates to true. A condition can add constraints based on attributes of the request, the resource, or both. To learn which resources support conditions in their IAM policies, see the IAM documentation. JSON example: { "bindings": [ { "role": "roles/resourcemanager.organizationAdmin", "members": [ "user:mike@example.com", "group:admins@example.com", "domain:google.com", "serviceAccount:my-project-id@appspot.gserviceaccount.com" ] }, { "role": "roles/resourcemanager.organizationViewer", "members": [ "user:eve@example.com" ], "condition": { "title": "expirable access", "description": "Does not grant access after Sep 2020", "expression": "request.time < timestamp('2020-10-01T00:00:00.000Z')", } } ], "etag": "BwWWja0YfJA=", "version": 3 } YAML example: bindings: - members: - user:mike@example.com - group:admins@example.com - domain:google.com - serviceAccount:my-project-id@appspot.gserviceaccount.com role: roles/resourcemanager.organizationAdmin - members: - user:eve@example.com role: roles/resourcemanager.organizationViewer condition: title: expirable access description: Does not grant access after Sep 2020 expression: request.time < timestamp('2020-10-01T00:00:00.000Z') - etag: BwWWja0YfJA= - version: 3 For a description of IAM and its features, see the IAM documentation.

ProjectConfigCreateCall
Creates a new RuntimeConfig resource. The configuration name must be unique within project.

ProjectConfigDeleteCall
Deletes a RuntimeConfig resource.

ProjectConfigGetCall
Gets information about a RuntimeConfig resource.

ProjectConfigGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

ProjectConfigListCall
Lists all the RuntimeConfig resources within project.

ProjectConfigOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectConfigOperationTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

ProjectConfigSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

ProjectConfigTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

ProjectConfigUpdateCall
Updates a RuntimeConfig resource. The configuration must exist beforehand.

ProjectConfigVariableCreateCall
Creates a variable within the given configuration. You cannot create a variable with a name that is a prefix of an existing variable name, or a name that has an existing variable name as a prefix. To learn more about creating a variable, read the Setting and Getting Data documentation.

ProjectConfigVariableDeleteCall
Deletes a variable or multiple variables. If you specify a variable name, then that variable is deleted. If you specify a prefix and recursive is true, then all variables with that prefix are deleted. You must set a recursive to true if you delete variables by prefix.

ProjectConfigVariableGetCall
Gets information about a single variable.

ProjectConfigVariableListCall
Lists variables within given a configuration, matching any provided filters. This only lists variable names, not the values, unless return_values is true, in which case only variables that user has IAM permission to GetVariable will be returned.

ProjectConfigVariableTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

ProjectConfigVariableUpdateCall
Updates an existing variable with a new value.

ProjectConfigVariableWatchCall
Watches a specific variable and waits for a change in the variable's value. When there is a change, this method returns the new value or times out. If a variable is deleted while being watched, the variableState state is set to DELETED and the method returns the last known variable value. If you set the deadline for watching to a larger value than internal timeout (60 seconds), the current variable value is returned and the variableState will be VARIABLE_STATE_UNSPECIFIED. To learn more about creating a watcher, read the Watching a Variable for Changes documentation.

ProjectConfigWaiterCreateCall
Creates a Waiter resource. This operation returns a long-running Operation resource which can be polled for completion. However, a waiter with the given name will exist (and can be retrieved) prior to the operation completing. If the operation fails, the failed Waiter resource will still exist and must be deleted prior to subsequent creation attempts.

ProjectConfigWaiterDeleteCall
Deletes the waiter with the specified name.

ProjectConfigWaiterGetCall
Gets information about a single waiter.

ProjectConfigWaiterListCall
List waiters within the given configuration.

ProjectConfigWaiterTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudRuntimeConfig hub.

RuntimeConfig
A RuntimeConfig resource is the primary resource in the Cloud RuntimeConfig service. A RuntimeConfig resource consists of metadata and a hierarchy of variables.

SetIamPolicyRequest
Request message for SetIamPolicy method.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

TestIamPermissionsRequest
Request message for TestIamPermissions method.

TestIamPermissionsResponse
Response message for TestIamPermissions method.

Variable
Describes a single variable within a RuntimeConfig resource. The name denotes the hierarchical variable name. For example, ports/serving_port is a valid variable name. The variable value is an opaque string and only leaf variables can have values (that is, variables that do not have any child variables).

Waiter
A Waiter resource waits for some end condition within a RuntimeConfig resource to be met before it returns. For example, assume you have a distributed system where each node writes to a Variable resource indicating the node's readiness as part of the startup process. You then configure a Waiter resource with the success condition set to wait until some number of nodes have checked in. Afterwards, your application runs some arbitrary code after the condition has been met and the waiter returns successfully. Once created, a Waiter resource is immutable. To learn more about using waiters, read the Creating a Waiter documentation.

WatchVariableRequest
Request for the WatchVariable() method.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.